common: Clean up fdwalk duplication, move to Linux close_range API
Linux 5.9 finally added a close_range() call [1]. glibc wraps this [2],
but not in a stable release yet. Let's move our code to be ready to use
it once it is available. Add a new common cockpit_close_range() function
which offers the same API, and use it everywhere.

Remove cockpit_unix_fd_close_*() from cockpitunixfd.h. It does not fit
there, as (1) that file otherwise contains backports of [2] and will
eventually disappear, and (2) we also want to use close_range() from
programs that don't include glib.

This also gets rid of some fdwalk() implementation duplication.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=278a5fbaed89dacd04e9d
[2] https://sourceware.org/git/?p=glibc.git;a=commit;h=286286283e9bdc
[3] https://developer.gnome.org/glib/stable/glib-UNIX-specific-utilities-and-integration.html
