+++ src/common/Makefile-common.am
@@ -56,6 +56,8 @@ noinst_LIBRARIES += libcockpit-common.a
 libcockpit_common_a_SOURCES = \
 	src/common/cockpitchannel.c \
 	src/common/cockpitchannel.h \
+	src/common/cockpitcloserange.c \
+	src/common/cockpitcloserange.h \
 	src/common/cockpitcontrolmessages.c \
 	src/common/cockpitcontrolmessages.h \
 	src/common/cockpiterror.h src/common/cockpiterror.c \
