+++ src/common/cockpitunixfd.c
@@ -23,10 +23,6 @@
 
 #include <glib-unix.h>
 
-#include <sys/resource.h>
-
-#include <dirent.h>
-#include <errno.h>
 #include <stdlib.h>
 #include <unistd.h>
 
@@ -120,137 +116,4 @@ cockpit_unix_fd_add_full (gint priority,
   g_source_unref (source);
 
   return ret;
-}
-
-typedef struct {
-  int from;
-  int except;
-  int until;
-} CloseAll;
-
-static int
-closefd (void *data,
-         gint fd)
-{
-  CloseAll *ca = data;
-
-  if (fd >= ca->from && fd != ca->except && fd < ca->until)
-    {
-      while (close (fd) < 0)
-        {
-          if (errno == EAGAIN || errno == EINTR)
-            continue;
-          if (errno == EBADF || errno == EINVAL)
-            break;
-          g_critical ("couldn't close fd in child process: %s", g_strerror (errno));
-          return -1;
-        }
-    }
-
-  return 0;
-}
-
-#ifndef HAVE_FDWALK
-static int
-fdwalk (int (*cb)(void *data, int fd), void *data)
-{
-  gint open_max;
-  gint fd;
-  gint res = 0;
-
-  struct rlimit rl;
-
-#ifdef __linux__
-  DIR *d;
-
-  if ((d = opendir("/proc/self/fd"))) {
-      struct dirent *de;
-
-      while ((de = readdir(d))) {
-          glong l;
-          gchar *e = NULL;
-
-          if (de->d_name[0] == '.')
-              continue;
-
-          errno = 0;
-          l = strtol(de->d_name, &e, 10);
-          if (errno != 0 || !e || *e)
-              continue;
-
-          fd = (gint) l;
-
-          if ((glong) fd != l)
-              continue;
-
-          if (fd == dirfd(d))
-              continue;
-
-          if ((res = cb (data, fd)) != 0)
-              break;
-        }
-
-      closedir(d);
-      return res;
-  }
-
-  /* If /proc is not mounted or not accessible we fall back to the old
-   * rlimit trick */
-
-#endif
-
-  if (getrlimit(RLIMIT_NOFILE, &rl) == 0 && rl.rlim_max != RLIM_INFINITY)
-      open_max = rl.rlim_max;
-  else
-      open_max = sysconf (_SC_OPEN_MAX);
-
-  for (fd = 0; fd < open_max; fd++)
-      if ((res = cb (data, fd)) != 0)
-          break;
-
-  return res;
-}
-
-#endif /* HAVE_FDWALK */
-
-/**
- * cockpit_unix_fd_close_all:
- * @from: minimum FD to close, or -1
- * @except: an FD to leave open, or -1
- *
- * Close all open file descriptors starting from @from
- * and skipping @except.
- *
- * Will set errno if a failure happens.
- *
- * Returns: zero if successful, -1 if not
- */
-int
-cockpit_unix_fd_close_all (int from,
-                           int except)
-{
-  CloseAll ca = { from, except, G_MAXINT };
-  return fdwalk (closefd, &ca);
-}
-
-/**
- * cockpit_unix_fd_close_until:
- * @from: minimum FD to close, or -1
- * @except: an FD to leave open, or -1
- * @until: stop closing fds when this number is hit.
- *
- * Close all open file descriptors starting from @from
- * and skipping @except up to but not including @until.
- *
- * Will set errno if a failure happens.
- *
- * Returns: zero if successful, -1 if not
- */
-int
-cockpit_unix_fd_close_until (int from,
-                             int except,
-                             int until)
-{
-  CloseAll ca = { from, except, until };
-  return fdwalk (closefd, &ca);
 }
