+++ src/ws/session-utils.c
@@ -19,15 +19,14 @@
 
 #include "session-utils.h"
 
+#include "common/cockpitcloserange.h"
 #include "common/cockpitframe.h"
 #include "common/cockpitjsonprint.h"
 
-#include <dirent.h>
 #include <fcntl.h>
 #include <stdarg.h>
 #include <stdlib.h>
 #include <sys/param.h>
-#include <sys/resource.h>
 #include <time.h>
 #include <utmp.h>
 
@@ -556,92 +555,6 @@ utmp_log (int login,
     }
 }
 
-int
-closefd (void *data,
-         int fd)
-{
-  int *from = data;
-  if (fd >= *from)
-    {
-      while (close (fd) < 0)
-        {
-          if (errno == EAGAIN || errno == EINTR)
-            continue;
-          if (errno == EBADF || errno == EINVAL)
-            break;
-          warnx ("couldn't close fd in bridge process: %m");
-          return -1;
-        }
-    }
-
-  return 0;
-}
-
-#ifndef HAVE_FDWALK
-
-int
-fdwalk (int (*cb)(void *data, int fd),
-        void *data)
-{
-  int open_max;
-  int fd;
-  int res = 0;
-
-  struct rlimit rl;
-
-#ifdef __linux__
-  DIR *d;
-
-  if ((d = opendir ("/proc/self/fd"))) {
-      struct dirent *de;
-
-      while ((de = readdir (d))) {
-          long l;
-          char *e = NULL;
-
-          if (de->d_name[0] == '.')
-              continue;
-
-          errno = 0;
-          l = strtol (de->d_name, &e, 10);
-          if (errno != 0 || !e || *e)
-              continue;
-
-          fd = (int) l;
-
-          if ((long) fd != l)
-              continue;
-
-          if (fd == dirfd (d))
-              continue;
-
-          if ((res = cb (data, fd)) != 0)
-              break;
-        }
-
-      closedir (d);
-      return res;
-  }
-
-  /* If /proc is not mounted or not accessible we fall back to the old
-   * rlimit trick */
-
-#endif
-
-  if (getrlimit (RLIMIT_NOFILE, &rl) == 0 && rl.rlim_max != RLIM_INFINITY)
-      open_max = rl.rlim_max;
-  else
-      open_max = sysconf (_SC_OPEN_MAX);
-
-  for (fd = 0; fd < open_max; fd++)
-      if ((res = cb (data, fd)) != 0)
-          break;
-
-  return res;
-}
-
-#endif /* HAVE_FDWALK */
-
 void
 btmp_log (const char *username,
           const char *rhost)
@@ -814,7 +727,7 @@ fd_remap (const int *remap_fds,
         abort_with_message ("dup2(%d, %d) failed: %m", fds[i], i);
 
   /* close everything else */
-  if (fdwalk (closefd, &n_remap_fds) < 0)
+  if (cockpit_close_range (n_remap_fds, INT_MAX, 0) < 0)
     abort_with_message ("couldn't close all file descriptors");
 }
