+++ src/ws/session-utils.h
@@ -63,10 +63,6 @@ void build_string (char **buf, size_t *size, const char *str, size_t len);
 void authorize_logger (const char *data);
 void save_environment (void);
 void pass_to_child (int signo);
-#ifndef HAVE_FDWALK
-int fdwalk (int (*cb)(void *data, int fd), void *data);
-#endif
-int closefd (void *data, int fd);
 void utmp_log (int login, const char *rhost, FILE *messages);
 void btmp_log (const char *username, const char *rhost);
